game_server/requirements.txt
python-dotenv==0.19.2
pytest-asyncio~=0.21
requests==2.26.0
unittest-xml-reporting==3.0.4
